+++ src/error.rs
@@ -14,10 +14,9 @@
 // You should have received a copy of the GNU General Public License along with
 // this program. If not, see https://www.gnu.org/licenses/.
 
-use std::{
-    fmt::{self, Display, Formatter},
-    io::{self, Write},
-};
+use std::fmt::{Display, Formatter};
+use std::io::{IsTerminal, Write};
+use std::{fmt, io};
 use termcolor::{BufferWriter, Color, ColorChoice, ColorSpec, WriteColor};
 
 pub enum Warning {
@@ -47,7 +46,7 @@ pub enum ErrorKind {
 impl ErrorKind {
     /// Colorize warning|error output.
     pub fn colorize(&self) -> io::Result<()> {
-        let color_choice = match atty::is(atty::Stream::Stderr) {
+        let color_choice = match io::stderr().is_terminal() {
             true => ColorChoice::Auto,
             false => ColorChoice::Never,
         };
@@ -64,9 +63,9 @@ impl ErrorKind {
         let mut buffer = writer.buffer();
 
         buffer.set_color(ColorSpec::new().set_fg(color).set_bold(true))?;
-        write!(&mut buffer, "{}: ", prefix)?;
+        write!(&mut buffer, "{prefix}: ")?;
         buffer.reset()?;
-        writeln!(&mut buffer, "{}", self)?;
+        writeln!(&mut buffer, "{self}")?;
 
         writer.print(&buffer)
     }
@@ -74,21 +73,27 @@ impl ErrorKind {
 
 impl Display for ErrorKind {
     /// Print colored error message, but ONLY on Stderr stream.
+    #[rustfmt::skip]
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
         match self {
             Self::Warning(warning) => match warning {
-                Warning::UnexpectedScaleOpt => write!(
-                    f,
-                    "ignoring `scale` CLI option with non-raster output format"
-                ),
+                Warning::UnexpectedScaleOpt => {
+                    write!(f, "Ignoring `scale` CLI option with non-raster output format")
+                },
             },
             Self::Error(error) => match error {
-                Error::QrCodeErr(msg) => write!(f, "unable to generate QR code: {msg}"),
-                Error::InvalidOutputExt => write!(f, "invalid output file extension"),
-                Error::SvgOutputErr(msg) => write!(f, "unable to write SVG output file: {msg}"),
+                Error::QrCodeErr(msg) => {
+                    write!(f, "Failed to generate QR code: {msg}")
+                },
+                Error::InvalidOutputExt => {
+                    write!(f, "Invalid output file extension, expected one of jpeg, jpg, png, svg")
+                },
+                Error::SvgOutputErr(msg) => {
+                    write!(f, "Failed to write SVG output file: {msg}")
+                },
                 Error::RasterOutputErr(msg) => {
-                    write!(f, "unable to write raster image file: {msg}")
-                }
+                    write!(f, "Failed to write raster image file: {msg}")
+                },
             },
         }
     }